+++ src/KhsCI/Service/PullRequest/GitHubClient.php
@@ -7,6 +7,11 @@ namespace KhsCI\Service\PullRequest;
 use Exception;
 use KhsCI\Service\CICommon;
 
+/**
+ * Class GitHubClient
+ *
+ * @see https://developer.github.com/v3/pulls/
+ */
 class GitHubClient
 {
     use CICommon;
@@ -207,7 +212,7 @@ class GitHubClient
             return false;
         }
 
-        return false;
+        throw new Exception('pull_request is merged error', 500);
     }
 
     /**
@@ -217,8 +222,7 @@ class GitHubClient
      * @param string $commit_title
      * @param string $commit_message
      * @param string $sha
-     * @param bool   $rebase
-     * @param bool   $squash
+     * @param int    $merge_method
      *
      * @return bool|mixed
      *
@@ -228,26 +232,27 @@ class GitHubClient
                           string $repo_name,
                           int $pr_num,
                           string $commit_title,
-                          string $commit_message,
+                          ?string $commit_message,
                           string $sha,
-                          bool $rebase = false,
-                          bool $squash = false)
+                          int $merge_method)
     {
-        if ($rebase && $squash) {
-            throw new Exception('', 500);
+        switch ($merge_method) {
+            case 1:
+                $merge_method = 'merge';
+                break;
+            case 2:
+                $merge_method = 'squash';
+                break;
+            case 3:
+                $merge_method = 'rebase';
+                break;
         }
 
-        $merge_method = false;
-
-        $rebase && $merge_method = 'rebase';
-
-        $squash && $merge_method = 'squash';
-
-        $url = $this->api_url.implode('/', ['/repos', $username, $repo_name, '/pulls', $pr_num, 'merge']);
+        $url = $this->api_url.implode('/', ['/repos', $username, $repo_name, 'pulls', $pr_num, 'merge']);
 
         $data = [
             'commit_title' => $commit_title,
-            'commit_message' => $commit_message,
+            'commit_message' => $commit_message ?? '',
             'sha' => $sha,
             'merge_method' => $merge_method,
         ];
@@ -261,13 +266,13 @@ class GitHubClient
         }
 
         if (405 === $http_return_code) {
-            return $output;
+            throw new Exception('merge cannot be performed', 405);
         }
 
         if (409 === $http_return_code) {
-            return $output;
+            throw new Exception('sha was provided and pull request head did not match', 409);
         }
 
-        return $output;
+        throw new Exception($output, $http_return_code);
     }
 }